requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Keras==2.4.3
line-bot-sdk==1.17.0
MarkupSafe==1.1.1
numpy==1.19.3
Pillow==8.0.1
PyYAML==5.3.1
requests==2.24.0
scipy==1.5.3
six==1.15.0
urllib3==1.25.11
Werkzeug==1.0.1
